apps/admin/pyproject.toml
[project]
name = "ras-admin"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "psycopg[binary]>=3.1",
    "pydantic-settings>=2.0",
    "jinja2",
    "python-multipart>=0.0.9",
    "httpx",
    "pyjwt>=2.0",
]

[project.optional-dependencies]
cloud = ["boto3>=1.34"]

[project.scripts]
ras-admin = "ras_admin.api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ras_admin"]
